@@ -1,13 +1,9 @@
 name: bootjack
 version: 1.0.1
-authors:
-- Simon Pai <simonpai@rikulo.org>
-- Jimmy Shiau <jimmyshiau@potix.com>
-- Tom Yeh <tomyeh@rikulo.org>
 description: Bootjack is a porting of Twitter Bootstrap 3.0.x in Dart.
 homepage: https://github.com/rikulo/bootjack
 documentation: http://blog.rikulo.org/posts/2013/May/General/bootjack-and-dquery/ 
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: '>=2.14.0 <3.0.0'
 dependencies:
   dquery: any
