@@ -1,7 +1,7 @@
 part of bootjack;
 
-Map<K, V> _copy<K, V>(Map<K, V> a, Map<K, V> b) {
-  final Map<K, V> c = new Map<K, V>.from(a);
+Map<K, V> _copy<K, V>(Map<K, V> a, Map<K, V>? b) {
+  final c = <K, V>{...a};
   if (b != null)
     b.forEach((key, value) => c[key] = value);
   return c;
